:iphone: fix responsiveness of the qr-button
