+++ client/src/constants.js
@@ -17,38 +17,3 @@ export const toastFail = {
     "--toastBarBackground": "#F0634D",
   },
 };
-// qr functions
-export function QRdownload(e) {
-  var myDiv = document.getElementById(API.KZILLA_URL + data.shortCode);
-  var myImage = myDiv.children[1];
-  var hr = document.getElementById(data.shortCode);
-  hr.href = myImage.src;
-  return false;
-}
-
-//Opening and closing the edit and QR modals.
-
-export function showEditor(e) {
-  var editor = this.parentElement;
-  window.$(editor).find(".kz-modalId").removeClass("d-none");
-}
-
-export function showQREditor(e) {
-  console.log("OUTSIDE FORMSTRUCTURE");
-  var editor = this.parentElement;
-  window.$(editor).find(".kz-qr-modalId").removeClass("d-none");
-}
-
-export function hideEditor() {
-  var editor = document.querySelectorAll(".kz-modalId");
-  editor.forEach(function (e) {
-    window.$(e).addClass("d-none");
-  });
-}
-
-export function hideQREditor() {
-  var editor = document.querySelectorAll(".kz-qr-modalId");
-  editor.forEach(function (e) {
-    window.$(e).addClass("d-none");
-  });
-}
